Don't override TOTAL_MEMORY for emscripten compiled version
Trying to allocate such a large amount of memory is causing an exception
on mobile devices. Instead, we should limit the maximum number of
iterations to ensure we don't use too much memory (on all devices I've
tested we should not be able to run enough iterations to run out of
memory anyway in the 1 second allocated for moves).